refactor: split renderer.rs into renderer + widgets + hit_test
renderer.rs (1622 lines) → three focused files:
- renderer.rs (418) — DrawCtx, draw_scene orchestrator, half-block flush,
  terminal lifecycle
- widgets.rs (842) — all ratatui paint_* fns, TickerQueue, build_status_summary
- hit_test.rs (422) — hit_test_agent/furniture/coffee_machine/from_tui

Re-exports in renderer.rs maintain backwards compatibility for all callers.
